--- src/tests/test_checkReplies.py
@@ -1,46 +1,0 @@
-# Generated by Selenium IDE
-import pytest
-import time
-import json
-from selenium import webdriver
-from selenium.webdriver.common.by import By
-from selenium.webdriver.common.action_chains import ActionChains
-from selenium.webdriver.support import expected_conditions
-from selenium.webdriver.support.wait import WebDriverWait
-from selenium.webdriver.common.keys import Keys
-from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
-
-
-class TestCheckReplies:
-    def setup_method(self, method):
-        self.driver = webdriver.Chrome()
-        self.vars = {}
-
-    def teardown_method(self, method):
-        self.driver.quit()
-
-    def test_checkReplies(self):
-        self.driver.get("http://127.0.0.1:5000")
-        self.driver.set_window_size(1245, 1040)
-        self.driver.find_element(By.LINK_TEXT, "Browse Posts").click()
-        self.driver.find_element(
-            By.CSS_SELECTOR, ".card:nth-child(4) .btn-info"
-        ).click()
-        assert (
-            self.driver.find_element(By.CSS_SELECTOR, "#replies-1 .card-subtitle").text
-            == "Reply by Anonymous"
-        )
-        self.driver.find_element(
-            By.CSS_SELECTOR, ".card:nth-child(6) .btn-info"
-        ).click()
-        assert (
-            self.driver.find_element(By.CSS_SELECTOR, "#replies-2 .card-subtitle").text
-            == "Reply by Anonymous"
-        )
-        self.driver.find_element(
-            By.CSS_SELECTOR, ".card:nth-child(8) .btn-info"
-        ).click()
-        assert (
-            self.driver.find_element(By.CSS_SELECTOR, "#replies-3 .card-subtitle").text
-            == "Reply by Anonymous"
-        )
