+++ src/tests/test_selenium.py
@@ -45,7 +45,7 @@ class Test_login:
         self.driver.find_element(By.ID, "login-password").send_keys("password123")
         self.driver.find_element(By.CSS_SELECTOR, ".btn:nth-child(6)").click()
 
-        self.driver.implicitly_wait(1)
+        self.driver.implicitly_wait(10)
 
         self.driver.find_element(By.CSS_SELECTOR, ".alert-success").click()
 
@@ -63,12 +63,6 @@ class Test_createPost:
         self.driver.get("http://127.0.0.1:5000")
         self.driver.set_window_size(1245, 1040)
         self.driver.find_element(By.LINK_TEXT, "Log In").click()
-        element = self.driver.find_element(By.LINK_TEXT, "Log In")
-        actions = ActionChains(self.driver)
-        actions.move_to_element(element).perform()
-        element = self.driver.find_element(By.CSS_SELECTOR, "body")
-        actions = ActionChains(self.driver)
-        actions.move_to_element(element).perform()
         self.driver.find_element(By.ID, "login-email").click()
         self.driver.find_element(By.ID, "login-email").send_keys("alice@example.com")
         self.driver.find_element(By.ID, "login-password").click()
@@ -77,8 +71,10 @@ class Test_createPost:
 
         self.driver.implicitly_wait(10)
 
-        elements = self.driver.find_elements(By.CSS_SELECTOR, ".alert-success")
-        assert len(elements) > 0
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "Successfully logged in\n×"
+        )
         self.driver.find_element(
             By.CSS_SELECTOR, ".col-md-3:nth-child(2) .card-title"
         ).click()
@@ -90,9 +86,154 @@ class Test_createPost:
 
         self.driver.implicitly_wait(10)
 
-        elements = self.driver.find_elements(By.CSS_SELECTOR, ".alert-success")
-        assert len(elements) > 0
-        elements = self.driver.find_elements(
-            By.CSS_SELECTOR, ".card:nth-child(4) .card-title"
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "Post created successfully!\n×"
         )
-        assert len(elements) > 0
+
+        self.driver.implicitly_wait(10)
+
+        assert (
+            self.driver.find_element(
+                By.CSS_SELECTOR, ".card:nth-child(4) .card-title"
+            ).text
+            == "Selenium Test"
+        )
+
+
+class Test_checkReplies:
+    def setup_method(self, method):
+        self.driver = webdriver.Chrome()
+        self.vars = {}
+
+    def teardown_method(self, method):
+        self.driver.quit()
+        reset_database()
+
+    def test_checkReplies(self):
+        self.driver.get("http://127.0.0.1:5000")
+        self.driver.set_window_size(1245, 1040)
+        self.driver.find_element(By.LINK_TEXT, "Browse Posts").click()
+        self.driver.find_element(
+            By.CSS_SELECTOR, ".card:nth-child(4) .btn-info"
+        ).click()
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, "#replies-1 .card-subtitle").text
+            == "Reply by Anonymous"
+        )
+        self.driver.find_element(
+            By.CSS_SELECTOR, ".card:nth-child(6) .btn-info"
+        ).click()
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, "#replies-2 .card-subtitle").text
+            == "Reply by Anonymous"
+        )
+        self.driver.find_element(
+            By.CSS_SELECTOR, ".card:nth-child(8) .btn-info"
+        ).click()
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, "#replies-3 .card-subtitle").text
+            == "Reply by Anonymous"
+        )
+
+class Test_passwordChange:
+    def setup_method(self, method):
+        self.driver = webdriver.Chrome()
+        self.vars = {}
+        reset_database()
+
+    def teardown_method(self, method):
+        self.driver.quit()
+        reset_database()
+
+    def test_passwordChange(self):
+        self.driver.get("http://127.0.0.1:5000")
+        self.driver.set_window_size(1285, 1039)
+        self.driver.find_element(By.CSS_SELECTOR, "body").click()
+        self.driver.find_element(By.LINK_TEXT, "Log In").click()
+        element = self.driver.find_element(By.LINK_TEXT, "Log In")
+        actions = ActionChains(self.driver)
+        actions.move_to_element(element).perform()
+        element = self.driver.find_element(By.CSS_SELECTOR, "body")
+        actions = ActionChains(self.driver)
+        actions.move_to_element(element).perform()
+        self.driver.find_element(By.CSS_SELECTOR, "#loginModal .modal-header").click()
+        self.driver.find_element(By.ID, "login-email").click()
+        self.driver.find_element(By.ID, "login-email").send_keys("alice@example.com")
+        self.driver.find_element(By.ID, "login-password").click()
+        self.driver.find_element(By.ID, "login-password").send_keys("password123")
+        self.driver.find_element(By.CSS_SELECTOR, ".btn:nth-child(6)").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "Successfully logged in\n×"
+        )
+        self.driver.find_element(By.LINK_TEXT, "Alice Johnson").click()
+        self.driver.find_element(By.ID, "changePasswordButton").click()
+        self.driver.find_element(By.ID, "current_password").click()
+        self.driver.find_element(By.ID, "current_password").send_keys("password123")
+        self.driver.find_element(By.ID, "new_password").click()
+        self.driver.find_element(By.ID, "new_password").send_keys("password1234")
+        self.driver.find_element(By.ID, "confirm_password").click()
+        self.driver.find_element(By.ID, "confirm_password").send_keys("password1234")
+        self.driver.find_element(By.CSS_SELECTOR, "#passwordForm > .btn").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.ID, "passwordError").text
+            == "Password must contain at least one uppercase letter."
+        )
+        self.driver.find_element(By.ID, "new_password").clear()
+        self.driver.find_element(By.ID, "confirm_password").clear()
+        self.driver.find_element(By.ID, "new_password").click()
+        self.driver.find_element(By.ID, "new_password").send_keys("Password1234")
+        self.driver.find_element(By.ID, "confirm_password").click()
+        self.driver.find_element(By.ID, "confirm_password").send_keys("Password1234")
+        self.driver.find_element(By.CSS_SELECTOR, "#passwordForm > .btn").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.ID, "passwordError").text
+            == "Password must contain at least one special character."
+        )
+        self.driver.find_element(By.ID, "new_password").clear()
+        self.driver.find_element(By.ID, "confirm_password").clear()
+        self.driver.find_element(By.ID, "new_password").click()
+        self.driver.find_element(By.ID, "new_password").send_keys("Password1234!")
+        self.driver.find_element(By.ID, "confirm_password").click()
+        self.driver.find_element(By.ID, "confirm_password").send_keys("Password1234!")
+        self.driver.find_element(By.CSS_SELECTOR, "#passwordForm > .btn").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "Password successfully updated\n×"
+        )
+        self.driver.find_element(By.LINK_TEXT, "Log Out").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "You have been logged out.\n×"
+        )
+        self.driver.find_element(By.LINK_TEXT, "Log In").click()
+        self.driver.find_element(By.ID, "login-email").click()
+        self.driver.find_element(By.ID, "login-email").send_keys("alice@example.com")
+        self.driver.find_element(By.ID, "login-password").click()
+        self.driver.find_element(By.ID, "login-password").send_keys("Password1234!")
+        self.driver.find_element(By.CSS_SELECTOR, ".btn:nth-child(6)").click()
+
+        time.sleep(0.5)
+
+        assert (
+            self.driver.find_element(By.CSS_SELECTOR, ".alert-success").text
+            == "Successfully logged in\n×"
+        )
+
+
